Create button for creating div elements
Included a button that calls a prompt for the user to input the desired
dimension for the square grid. Used console.log(typeof dimension) to see
if the value entered is a string or number. Since it was a string, I
used parseInt(dimension) to convert the string into a number to later
generate the desired grid of div elements.