Edit function to create div elements
Moved the createContainer() function inside the event listener for the
button. Defined the dimensions for rows and columns as "i" and "j"
within for loops. Removed element = "div" + i and console.log because I
no longer needed to track each div. Changed the display from "flex" to
"grid" and defined the style of the rows as "repeat(dimension, 1fr).
Although the function properly creates the correct number of div
elements, the grid format is incorrect; the function creates a column of
div elements instead of a grid of rows and columns.
}