Fix grid sizing issue
Was able to turn all the rows of divs into a grid based on the user's
numerical input. I needed to use JavaScript to dynamically resize the
grid using the variable dimension. At first I set both
container.style.gridTemplateColumns and container.style.gridTemplateRows
= "repeat(numberDimension, 1fr)", but this only returned rows of div
elements instead of a grid. The issue was fixed with
`repeat(${newDimension}, 1fr)' for both columns and rows.