Attempt to create button to make new grid
Wanted to fix the formatting issues with the grid container and div
elements when a new grid was created. Tried to create a new button that
would erase the old grid before making a new grid. The original button
became disabled. Could not recreate grid or fix the formatting issues.