Fix color-changing error
To get the div elements to change color when the mouse moved over them,
I had to move the 'mouseenter' and 'mouseleave' event listeners inside
the click event handler for the button. I changed the CSS display
setting to grid, removed the j variable and changed the for loop
condition to "i < dimension * dimension". Inside the button's click
event handler, I included "container.innerHTML = '';" to remove the grid
whenever the button is clicked.