Create rainbow button to use random colors
Edited index.html to include buttons "original" and "rainbow" that
changes the color of the div elements' background to different colors
when selected. The original button uses the green shade previously used.
The rainbox button should change the color of each div randomly, but
currently all divs are the same color. The rainbow button currently
switches to different colors randomly when it is clicked. This was done
with a "click" event listener that uses Math.floor(Math.random() *
array.length) with a defined array of colors. Each array element's
number is assigned a color in the function assignColor(). This function
was then used in both "mouseenter" and "mouseleave" event listeners to
assign the div elements' new background colors.