Fix mouse-related event handlers issue
The background color that was generated within the "mouseenter" event
handler was maintained in the "mouseleave" event handler by removing the
assignColor() function from the "mouseleave" event handler and assigning
randomColor = assignColor() so that this variable was used as the
background color in both event handlers. Also updated the look of the
Etch-a-Sketch.