Create new button and fix h1 element
The Marker button is a new option that lets the user draw on the sketch
pad in black, beginning with 10% opacity. When the user hovers over the
same div element repeatedly, the opacity will increase in 10% intervals
until the div element is completely black (100% opacity). Currently the
function only switches between 10% and 20% opacity. Also fixed the h1
element such that the text decoration is restyled by thickness and
offset, and left and right margins are set to auto and width is set to
max-content. This ensures the h1 element remains fixed when the browser
window resizes.