Fix markerButton and rename it
Renamed the marker button as pencil button. Instead of using if loop to
change the div elements' opacity by increments of 10%, I had to define
the opacity and divOpacity variables as 0.1 and rgba(0, 0, 0, 0.1)
respectively within the divs.forEach loop. Within the "mouseenter" event
handler, I had to use opacity += 0.1 and an if loop to increment the
opacity until it reached 100%. Then I had to define the background color
as divOpacity = `rgba(0, 0, 0, ${opacity}` for both "mouseenter" and
"mouseleave" event handlers.