Fix styling issues
Removed the borders from the div with class "grid-container". Replaced
the <p> element that had the text "Etch-a-Sketch" and the circular divs
with an image of how they look properly aligned, instead of trying to
use relative units to align each of them separately. Contained the
"grid-container" div inside another div displayed as a flexbox to create
the on-screen Etch-a-Sketch.