Create start button to reveal Etch-a-Sketch
Styled a start button with a rotating gradient animation and included a
class "none" to several elements such that only the title and start
button are visible in the beginnig. Applied an event listener to the
start button to remove the start button and show the on-screen Etch-a-
Sketch and pixelButton.